:recycle: refactor(cleanup): util함수를 한곳에 모아서 관리한다
