--- src/js/utils/utils.js
@@ -1,8 +1,0 @@
-const consoleErrorWithConditionalAlert = (error, errorNameForAlert) => {
-  console.error(error);
-  if (error.name === errorNameForAlert) {
-    alert(error.message);
-  }
-};
-
-export default consoleErrorWithConditionalAlert;
